// src/libs/data-access/klass/models/ClassificationVariantResource.ts
export interface ClassificationVariantResource {
    /**
     * 
     * @type {string}
     * @memberof ClassificationVariantResource
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof ClassificationVariantResource
     */
    id?: number;
    /**
     * 
     * @type {ContactPersonResource}
     * @memberof ClassificationVariantResource
     */
    contactPerson?: ContactPersonResource;
    /**
     * 
     * @type {string}
     * @memberof ClassificationVariantResource
     */
    owningSection?: string;
    /**
     * 
     * @type {Date}
     * @memberof ClassificationVariantResource
     */
    lastModified?: Date;
    /**
     * 
     * @type {Array<string>}
     * @memberof ClassificationVariantResource
     */
    published?: Array<string>;
    /**
     * 
     * @type {Date}
     * @memberof ClassificationVariantResource
     */
    validFrom?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ClassificationVariantResource
     */
    validTo?: Date;
    /**
     * 
     * @type {string}
     * @memberof ClassificationVariantResource
     */
    introduction?: string;
    /**
     * 
     * @type {Array<CorrespondenceTableSummaryResource>}
     * @memberof ClassificationVariantResource
     */
    correspondenceTables?: Array<CorrespondenceTableSummaryResource>;
    /**
     * 
     * @type {Array<ChangelogResource>}
     * @memberof ClassificationVariantResource
     */
    changelogs?: Array<ChangelogResource>;
    /**
     * 
     * @type {Array<LevelResource>}
     * @memberof ClassificationVariantResource
     */
    levels?: Array<LevelResource>;
    /**
     * 
     * @type {Array<ClassificationItemResource>}
     * @memberof ClassificationVariantResource
     */
    classificationItems?: Array<ClassificationItemResource>;
    /**
     * 
     * @type {{ [key: string]: Link; }}
     * @memberof ClassificationVariantResource
     */
    links?: { [key: string]: Link; };
    /**
     * 
     * @type {ClassificationVariantSummaryResourceEmbedded}
     * @memberof ClassificationVariantResource
     */
    embedded?: ClassificationVariantSummaryResourceEmbedded;
    /**
     * 
     * @type {PageMetadata}
     * @memberof ClassificationVariantResource
     */
    page?: PageMetadata;
}